requirements.txt
python-dotenv==1.0.0
pymongo==3.12.3
dnspython==2.4.2
mongoengine==0.27.0
djongo==1.3.6
django==4.2.7
PyJWT==2.8.0
ecdsa==0.18.0